pipeline/requirements.txt
# Core LLM / Search
litellm>=1.30.0
tavily-python>=0.3.0
duckduckgo-search>=5.0.0

# Web scraping
beautifulsoup4>=4.12.0
httpx>=0.27.0

# Reddit discovery
praw>=7.7.0

# Data models & validation
pydantic>=2.0.0

# Retry / resilience
tenacity>=8.2.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
# 显式指定兼容 Python 3.14 的 numpy 版本（knowledge-storm 依赖 numpy==1.26.4 会报错）
numpy>=2.0.0